backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Async Database
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.0

# AI / LLM
openai>=1.60.0

# RAG / Embedding
chromadb>=0.6.0
sentence-transformers>=3.4.0

# SQL Security
sqlglot>=26.0.0

# Async Support
greenlet>=3.0.0

# Configuration
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# SSE
sse-starlette>=2.2.0

# Logging
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
